scripts/requirements.txt
azure-cognitiveservices-speech==1.32.1
python-dotenv==1.0.0